Split mc into header and code
